serene_chatbot/requirements.txt
# Serene Design Studio Chatbot Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Gemini AI
google-generativeai>=0.3.2

# Data Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# CORS and HTTP utilities
python-multipart>=0.0.6
